requirements/build.txt
# These must be installed before building mmdetection
torch>=1.3
pycocotools
matplotlib
mmcv==0.5.6
numpy
# need older pillow until torchvision is fixed
Pillow<=6.2.2
six
terminaltables
torch>=1.3
torchvision
albumentations==0.4.6
Shapely
tqdm